requirements.txt
requests
feedparser
streamlit
playwright
openai
google-generativeai
python-pptx
nest_asyncio
